feat: Updated state management from local storage to pinia
