+++ package.json
@@ -12,8 +12,10 @@
   },
   "dependencies": {
     "axios": "^1.7.5",
+    "franc": "^6.2.0",
     "markdown-it": "^14.1.0",
     "oh-vue-icons": "^1.0.0-rc3",
+    "pinia": "^2.2.2",
     "vue": "^3.4.29",
     "vue-router": "^4.3.3",
     "vue-spinner": "^1.0.4"
